+++ packages/core/src/tool/index.ts
@@ -10,8 +10,15 @@ export default class Tool {
   instance: LogicFlow;
   constructor(instance: LogicFlow) {
     this.instance = instance;
-    this.registerTool('textEdit', TextEdit);
-    this.registerTool('multipleSelect', MultipleSelect);
+    if (!this.isDisabledTool(TextEdit.toolName)) {
+      this.registerTool(TextEdit.toolName, TextEdit);
+    }
+    if (!this.isDisabledTool(MultipleSelect.toolName)) {
+      this.registerTool(MultipleSelect.toolName, MultipleSelect);
+    }
+  }
+  private isDisabledTool(toolName) {
+    return this.instance.options.disabledTools.indexOf(toolName) !== -1;
   }
   registerTool(name, component) {
     this.toolMap.set(name, component);