+++ packages/extension/scripts/webpack.config.base.js
@@ -19,6 +19,7 @@ const packagesEntry = {
   InsertNodeInPolyline: path.resolve(__dirname, '../src/insert-node-in-polyline/index.ts'),
   FlowPath: path.resolve(__dirname, '../src/tools/flow-path/index.ts'),
   AutoLayout: path.resolve(__dirname, '../src/tools/auto-layout/index.ts'),
+  // MindMap: path.resolve(__dirname, '../src/mindmap/index.ts'),
 };
 
 module.exports = {