+++ examples/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.20](https://github.com/didi/LogicFlow/compare/examples@0.7.19...examples@0.7.20) (2021-11-30)
+
+**Note:** Version bump only for package examples
+
+
+
+
+
+## [0.7.19](https://github.com/didi/LogicFlow/compare/examples@0.7.18...examples@0.7.19) (2021-11-26)
+
+**Note:** Version bump only for package examples
+
+
+
+
+
 ## [0.7.18](https://github.com/didi/LogicFlow/compare/examples@0.7.17...examples@0.7.18) (2021-11-23)
 
 **Note:** Version bump only for package examples